+++ client/src/features/theme/index.js
@@ -1,4 +1,3 @@
 export { ThemeProvider } from './components/ThemeProvider';
-export { default as ThemeToggleCard } from './components/ThemeToggleCard';
 export { default as useTheme } from './hooks/useTheme';
 export * from './utils/tokens';